Avoid mixing different integer types
- Remove unnecessary code at the end of createLinks() function.

+++ Model/MaxIV/R3CommonInc/R3Ring.h
@@ -46,7 +46,7 @@ class R3Ring :
 {
  private:
 
-  const size_t NInnerSurf;           ///< Number of inner surf
+  const int NInnerSurf;           ///< Number of inner surf
   
   double fullOuterRadius;          ///< full outer radius [beamline endpoint]
